Add matching functionality to MatchCommand
This is an intermediate implementation of MatchCommand, which
is executed to match a patient with Donors in the Donor pool.
Currently when this command is executed, it displays all donors which
matches the specified patient.

Currently in this implementation, a match happens when patient and donor's
blood types are compatible and they pass the tissue typing test.