Make getMatchList defensive
Previously, getMatchList() returns a reference to the actual
match list in ModelManager. This leaves the original copy of the match list
available for modification.

Currently, getMatchList() is modified so a copy of the list is provided.
The original match list will not be modified by any changes made to the copy.